+++ src/Checkout/Payment/Method/GooglePayHandler.php
@@ -17,6 +17,7 @@ use Swag\PayPal\Checkout\Card\Exception\GooglePayValidationFailedException;
 use Swag\PayPal\Checkout\Payment\Service\OrderExecuteService;
 use Swag\PayPal\Checkout\Payment\Service\OrderPatchService;
 use Swag\PayPal\Checkout\Payment\Service\TransactionDataService;
+use Swag\PayPal\Checkout\Payment\Service\VaultTokenService;
 use Swag\PayPal\RestApi\V2\Api\Order;
 use Swag\PayPal\RestApi\V2\Resource\OrderResource;
 use Swag\PayPal\Setting\Service\SettingsValidationServiceInterface;
@@ -35,9 +36,10 @@ class GooglePayHandler extends AbstractSyncAPMHandler
         TransactionDataService $transactionDataService,
         LoggerInterface $logger,
         OrderResource $orderResource,
+        VaultTokenService $vaultTokenService,
         private readonly CardValidatorInterface $googlePayValidator
     ) {
-        parent::__construct($settingsValidationService, $orderTransactionStateHandler, $orderExecuteService, $orderPatchService, $transactionDataService, $logger, $orderResource);
+        parent::__construct($settingsValidationService, $orderTransactionStateHandler, $orderExecuteService, $orderPatchService, $transactionDataService, $logger, $orderResource, $vaultTokenService);
     }
 
     protected function executeOrder(SyncPaymentTransactionStruct $transaction, Order $paypalOrder, SalesChannelContext $salesChannelContext): Order